- make it compile on x86_64 - find the AMD64: tags to find the hacked crap that will need fixing
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/branches/rev_1_11@1402 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/db.c
@@ -4742,7 +4742,7 @@ char *file_to_string(char *name)
     char           *buf;
     char           *curr;
     struct stat     stat;
-    int             size;
+    long            size;
     int             rd;
     char            tmp[256];
